Add database integration (#2)

+++ src/webhook/types/checks.rs
@@ -54,7 +54,7 @@ pub struct CheckRunOutput {
 
 #[derive(Debug, Deserialize)]
 pub struct CheckSuite {
-    pub id: u32,
+    pub id: u64,
     pub node_id: String,
     pub head_branch: String,
     pub head_sha: String,
@@ -68,7 +68,7 @@ pub struct CheckSuite {
 
 #[derive(Debug, Deserialize)]
 pub struct CheckRun {
-    pub id: u32,
+    pub id: u64,
     pub node_id: String,
     pub head_sha: String,
     pub external_id: String,
